Properly handle nested loops.
Previously "it" was always used for the variable in a loop,
and in a nested loop there was no way to distinguish the
value of the inner iteration from the value of the outer one.

Now we assign the iterated value to both "it" and to the
original variable name (e.g. "foo.bar").

This probably has a small negative performance impact.

Closes #15.