Remove HsYAML dependency.
Remove ToYAML, FromYAML instances for TemplateTarget.

[API change]

Motivation: these were needed when we used HsYAML in pandoc.
Now that we don't, they are pulling in an HsYAML dependency
we wouldn't otherwise have in pandoc.

I checked all the packages on Hackage that use doctemplates,
and none also depend on HsYAML, so I believe this won't be disruptive.